OMF-40 fix: 리뷰 반영하여 인스턴스 로직 수정 및 파일명 변경
